Wasm prototype handles memory

+++ compiler/tests_diff/test.sh
@@ -1,5 +1,30 @@
 #!/bin/bash
 
+# Parse command line
+VERBOSE=false
+ALL=false
+CLEAN=false
+
+while getopts "vac" opt; do
+  case $opt in
+    v)
+      VERBOSE=true
+      ;;
+    a)
+      ALL=true
+      ;;
+    c)
+      CLEAN=true
+      ;;
+    \?)
+      echo "Invalid option : -$OPTARG" >&2
+      exit 1
+      ;;
+  esac
+done
+
+shift $((OPTIND - 1))
+
 # Globals
 ROOT_DIR=$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" &> /dev/null && pwd)
 PARENT_DIR=$(dirname "$ROOT_DIR")
@@ -20,6 +45,9 @@ SEP="===========================================================================
 run_tests() {
   local FILES="$1"
   local SIZE="$2"
+
+  # Flag for non-verbose mode
+  error=false
 
   # Test on every *.jazz files
   for path in "$FILES"/*.jazz; do
@@ -43,38 +71,58 @@ run_tests() {
     f_wasm="$FILES/$base.wasm"
 
     # Separator
-    printf "\n%s\n\n" "$SEP"
+    if [ "$VERBOSE" = true ]; then
+      printf "\n%s\n\n" "$SEP"
+    elif [ "$error" = true ]; then
+      printf "\n"
+    fi
+
+    # Reset flag for non-verbose mode
+    error=false
 
     # Compile to x86-64
-    printf "Compile $base to x86-64...\n\n"
-    "$COMPILER" -arch x86-64 -pasm -nowarning "$f_jazz" > "$f_s" || continue
-    gcc -c "$f_c" -o "$f_main_o" || continue
-    gcc -c "$f_s" -o "$f_o" || continue
-    gcc -no-pie "$f_main_o" "$f_o" -o "$f_exe" || continue
+    if [ "$VERBOSE" = true ]; then
+      printf "Compile $base to x86-64...\n\n"
+    fi
+    "$COMPILER" -arch x86-64 -pasm -nowarning "$f_jazz" > "$f_s" || { error=true; continue; }
+    gcc -c "$f_c" -o "$f_main_o" || { error=true; continue; }
+    gcc -c "$f_s" -o "$f_o" || { error=true; continue; }
+    gcc -no-pie "$f_main_o" "$f_o" -o "$f_exe" || { error=true; continue; }
 
     # Compile to Wasm
-    printf "Compile $base to Wasm...\n\n"
-    "$COMPILER" -arch wasm -pasm -nowarning "$f_jazz" > "$f_wat" || continue
-    wat2wasm "$f_wat" -o "$f_wasm" || continue
+    if [ "$VERBOSE" = true ]; then
+      printf "Compile $base to Wasm...\n\n"
+    fi
+    "$COMPILER" -arch wasm -pasm -nowarning "$f_jazz" > "$f_wat" || { error=true; continue; }
+    wat2wasm "$f_wat" -o "$f_wasm" || { error=true; continue; }
 
     # Compare results
-    echo "Test $f_jazz :"
+    if [ "$VERBOSE" = true ]; then
+      echo "Test $f_jazz :"
+    fi
 
     for val in "${VALUES[@]}"; do
-      printf "[Input %4s] : " "$val"
-
       RESULT_X86=$("$f_exe" "$SIZE" "$val")
       RESULT_WASM=$(node "$f_js" "$f_wasm" "$SIZE" "$val")
 
       if [ "$RESULT_X86" == "$RESULT_WASM" ]; then
-        printf "${GREEN}%S${NC} %-10s\n" "OK" "$RESULT_WASM"
+        if [ "$VERBOSE" = true ]; then
+          printf "[Input %4s] : ${GREEN}%s${NC} %-10s\n" "$val" "OK" "$RESULT_WASM"
+        fi
       else
-        printf "${RED}%S${NC} X86: %-10s | WASM: %s\n" "ERROR" "$RESULT_X86" "$RESULT_WASM"
+        error=true
+        if [ "$VERBOSE" = true ]; then
+          printf "[Input %4s] : ${RED}%s${NC} X86: %-10s | WASM: %s\n" "$val" "ERROR" "$RESULT_X86" "$RESULT_WASM"
+        else
+          printf "[Input %4s] : ${RED}%s${NC} X86: %-10s | WASM: %-10s (%s)\n" "$val" "ERROR" "$RESULT_X86" "$RESULT_WASM" "$f_jazz"
+        fi
       fi
     done
 
     # Remove build files
-    rm -f "$f_s" "$f_main_o" "$f_o" "$f_exe" "$f_wat" "$f_wasm"
+    if [ "$CLEAN" = true ]; then
+      rm -f "$f_s" "$f_main_o" "$f_o" "$f_exe" "$f_wat" "$f_wasm"
+    fi
   done
 }
 
@@ -84,5 +132,15 @@ make -C "$PARENT_DIR"
 clear
 
 # Run the tests
-run_tests "$FILES_32" 32
-run_tests "$FILES_64" 64
+if [ "$ALL" = true ]; then
+  run_tests "$FILES_32" 32
+  run_tests "$FILES_64" 64
+else
+  run_tests "$FILES_32" 32
+fi
+
+# Remove build files
+if [ "$CLEAN" = true ]; then
+  rm -f "$FILES_32"/*.{s,o,exe,wat,wasm}
+  rm -f "$FILES_64"/*.{s,o,exe,wat,wasm}
+fi
